Update utils
The functions that held the changes to the table are now inside the table module where it shares all the table related functions
